M2-A: RGB565 framebuffer with imported RGB565 LUT
- LTDC RGB565, ImageWidth 1536 (stride 3072), DSI RGB888 frozen unchanged
- 16-bit pixel rendering/colors (0xF800/0x07E0/0x001F), gray ramp formula
- Imported RGB565 LUT active; BSP ARGB table dropped from board_lcd
  (single strong definition in Src/gfxmmu.c, extern in board_lcd)
- FB0/FB1 tightened to 384 KiB each (footprint 370,256 B + slack)
- Mapping verify 3/3 in RGB565, swap 104==104, 78.47 Hz, no new errors

+++ platform/display/framebuffer.h
@@ -12,8 +12,8 @@ extern "C" {
 #define FB_LOGICAL_HEIGHT 480UL
 #define FB_STRIDE_BYTES 3072UL
 
-extern uint32_t m_fb0_phys[];
-extern uint32_t m_fb1_phys[];
+extern uint16_t m_fb0_phys[];
+extern uint16_t m_fb1_phys[];
 
 extern volatile uint32_t g_fb_swap_submit_seq;
 extern volatile uint32_t g_fb_reload_done_seq;
